% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inftocor.r
\name{inftocor}
\alias{inftocor}
\title{Transform information fractions into correlation matrix}
\usage{
inftocor(ir = c(0.2, 0.5, 1))
}
\arguments{
\item{ir}{Numeric vector of the sequence of information fractions. All
elements should be between 0 and 1 with the last one being exactly 1.}
}
\value{
List with an element named \code{cor} for the correlation matrix.
}
\description{
\code{inftocor()} transforms information (fractions) into correlation matrix.
}
\examples{
inftocor(ir = c(0.2, 0.5, 1.0))
}
\author{
Xiaodong Luo
}
\concept{correlation matrix}
\concept{information fraction}
