\name{MultiGroupSequential-package}
\alias{MultiGroupSequential-package}
\alias{MultiGroupSequential}
\docType{package}
\title{
\packageTitle{MultiGroupSequential}
}
\description{
\packageDescription{MultiGroupSequential}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MultiGroupSequential}
\packageIndices{MultiGroupSequential}
}
\author{
\packageAuthor{MultiGroupSequential}

Maintainer: \packageMaintainer{MultiGroupSequential}
}
\references{
Xiaodong Luo & Hui Quan (2023) Some Multiplicity Adjustment Procedures for Clinical Trials with Sequential Design and Multiple Endpoints, Statistics in Biopharmaceutical Research, DOI: 10.1080/19466315.2023.2191989.

Willi Maurer & Frank Bretz (2013) Multiple Testing in Group Sequential Trials Using Graphical Approaches, Statistics in Biopharmaceutical Research, 5:4, 311-320, DOI: 10.1080/19466315.2013.807748
}
\keyword{multiplicity control}
\keyword{group-sequential}
\keyword{multiple endpoints}



