\name{inftocor}
\alias{inftocor}
\title{Transform information (fractions) into correlation matrix}
\description{This utility function transforms information (fractions) into correlation matrix}
\usage{
inftocor(ir=c(0.2,0.5,1))
}
\arguments{
  \item{ir}{information (fractions)}
}
\value{\item{cor}{correlation matrix}
}
\author{Xiaodong Luo}
\examples{
inftocor(ir=c(0.2,0.5,1))
}
\keyword{correlation matrix}
\keyword{information fraction}

